% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggregate_functions}
\alias{aggregate_functions}
\title{Aggregates compound values into indices at the function level}
\usage{
aggregate_functions(compound_table, functions)
}
\arguments{
\item{compound_table}{Compound table outputted by aggregate_compounds()}

\item{functions}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}
}
\value{
A function table aggregated at the compound level
}
\description{
Aggregates compound values into indices at the function level
}
\examples{
aggregate_functions(compound_table)
}
