% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expression.R
\name{compute_expression}
\alias{compute_expression}
\title{Calculates the overall fullness-expression of a metabolic pathway/module}
\usage{
compute_expression(definition, expression_table)
}
\arguments{
\item{definition}{Definition string of a given metabolic pathway/module}

\item{expression_table}{Table containing expression data per functon}
}
\value{
A fullness-expression value
}
\description{
Calculates the overall fullness-expression of a metabolic pathway/module
}
\examples{
compute_expression(definition,expression_table)
}
