% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fullness_community.R
\name{compute_fullness_community}
\alias{compute_fullness_community}
\title{Calculates the MCI of a metabolic pathway at the community level}
\usage{
compute_fullness_community(definition, abundance_vector)
}
\arguments{
\item{definition}{Definition string of a given metabolic pathway/module}

\item{abundance_vector}{Vector containing relative abundance data per identifier}
}
\value{
A MCI value
}
\description{
Calculates the MCI of a metabolic pathway at the community level
}
\examples{
compute_fullness(definition,present)
compute_fullness("K01580 (K13524,K07250,K00823,K16871) (K00135,K00139,K17761)",c("K01580","K00823","K16871"))
}
