% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dama_functions.R
\name{dama_functions}
\alias{dama_functions}
\title{Aggregates compound values into indices at the function level}
\usage{
dama_functions(compounds_table, functions_table, normalise = FALSE)
}
\arguments{
\item{normalise}{Whether to normalise the scale to the maximum values in the dataset. Default=FALSE}

\item{compound_table}{Compound table outputted by aggregate_compounds()}

\item{functions}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}
}
\value{
A function table aggregated at the compound level
}
\description{
Aggregates compound values into indices at the function level
}
\examples{
dama_functions(compound_table)
}
