% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_bin.R
\name{damma_binary}
\alias{damma_binary}
\title{Convert the fullness table of metabolic pathways/modules into a binary (present/absent) matrix}
\usage{
damma_binary(fullness_table, threshold = 0.9)
}
\arguments{
\item{fullness_table}{Table containing fullness scores of functions (X axis) per MAG (Y axis)}

\item{threshold}{Minimum fullness value to consider a function (pathway/module) to be present in a MAG}
}
\value{
A binary function matrix
}
\description{
Convert the fullness table of metabolic pathways/modules into a binary (present/absent) matrix
}
\examples{
damma_binary(fullness_table)
damma_binary(fullness_table,0.9)
damma_binary(fullness_table,threshold=0.9)
}
