% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_community.R
\name{damma_community}
\alias{damma_community}
\title{Generate community-level MCIs from microbial relative abundance data}
\usage{
damma_community(
  annotations,
  functions_table,
  abundance_table,
  fullness_table,
  genomes_completeness,
  genomecol,
  keggcol,
  eccol,
  pepcol
)
}
\arguments{
\item{annotations}{Table containing Genome identifiers and annotation codes}

\item{functions_table}{Table containing definitions and metadata of metabolic functions (included in DAMMA package)}

\item{abundance_table}{(Relative) abundance table with samples in columns and Genomes in rows. Required for computing sample-specific MCI values}

\item{fullness_table}{Pathway fullness table provided by damma() function}

\item{genomes_completeness}{Genome completeness table to correct MCIs for varying completeness values}

\item{genomecol}{Column index (number) of the annotations table containing the MAG identifiers}

\item{keggcol}{Column index(es) of the annotations table in which to search for KEGG KO annotations}

\item{eccol}{Column index(es) of the annotations table in which to search for Enzyme Commision (EC) annotations}

\item{pepcol}{Column index(es) of the annotations table in which to search for Peptidase annotations}
}
\value{
A pathway fullness vector (if no abundance data are provided) or matrix (if abundance data are provided)
}
\description{
Generate community-level MCIs from microbial relative abundance data
}
\examples{
damma_community(annotations,functions_table,abundance_table,fullness_table,genomecol,keggcol,eccol,pepcol)
damma_community(annotations,functions_table,genomecol,keggcol,eccol,pepcol)
}
