% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_community_expression.R
\name{damma_community_expression}
\alias{damma_community_expression}
\title{Generate community-level MCIs from microbial gene expression data}
\usage{
damma_community_expression(
  annotations,
  functions_table,
  expression,
  fullness_table_list,
  genecol,
  genomecol,
  keggcol,
  eccol,
  pepcol
)
}
\arguments{
\item{annotations}{Table containing Genome identifiers and annotation codes}

\item{functions_table}{Table containing definitions and metadata of metabolic functions (included in DAMMA package)}

\item{expression}{Gene expression data}

\item{fullness_table_list}{A list of pathway-expression matrices (one table per genome)}

\item{genecol}{Column index (number) of the annotations table containing the gene identifiers}

\item{genomecol}{Column index (number) of the annotations table containing the genome identifiers}

\item{keggcol}{Column index(es) of the annotations table in which to search for KEGG KO annotations}

\item{eccol}{Column index(es) of the annotations table in which to search for Enzyme Commision (EC) annotations}

\item{pepcol}{Column index(es) of the annotations table in which to search for Peptidase annotations}
}
\value{
A pathway fullness vector (if no abundance data are provided) or matrix (if abundance data are provided)
}
\description{
Generate community-level MCIs from microbial gene expression data
}
\examples{
damma_community_expression(annotations,functions_table,expression,fullness_table_list,genecol,genomecol,keggcol,eccol,pepcol)
damma_community_expression(annotations=gene_annotations,functions_table,expression=gene_expression,genecol=1,genomecol=2,keggcol=9,eccol=c(10,19),pepcol=12)
}
