% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_compounds.R
\name{damma_compounds}
\alias{damma_compounds}
\title{Aggregates pathway-level MCIs into compound-level MCIs}
\usage{
damma_compounds(MCI_table, pathway_table)
}
\arguments{
\item{MCI_table}{Function distillation table outputted by damma() or damma_correction}

\item{pathway_table}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}
}
\value{
A MCI table aggregated at the compound level
}
\description{
Aggregates pathway-level MCIs into compound-level MCIs
}
\examples{
damma_compounds(compound_table,pathway_table)
}
