% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_correction.R
\name{damma_correction}
\alias{damma_correction}
\title{Models the relationship between function fullness and MAG completeness and applies a correction factor}
\usage{
damma_correction(fullness_table, genome_completeness, stats = TRUE)
}
\arguments{
\item{fullness_table}{Table containing fullness scores of functions (X axis) per MAG (Y axis)}

\item{genome_completeness}{A matrix containing MAG names (1st column) and completemess values (2nd column)}

\item{stats}{Whether to print correction statistics on screen or not. Default=TRUE}
}
\value{
A corrected fullness table
}
\description{
Models the relationship between function fullness and MAG completeness and applies a correction factor
}
\examples{
damma_correction(fullness_table, genome_completeness)
}
