% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_correlation.R
\name{damma_correlation}
\alias{damma_correlation}
\title{Calculate correlations between compounds and core functions}
\usage{
damma_correlation(fullness_table, method = "spearman", output = "plot")
}
\arguments{
\item{fullness_table}{Table containing fullness scores of functions (X axis) per MAG (Y axis)}

\item{method}{Correlation method to be used. Default="pearson"}

\item{output}{Whether to output a table or a plot. Default="plot"}
}
\value{
A correlation matrix or plot
}
\description{
Calculate correlations between compounds and core functions
}
\examples{
damma_cor(fullness_table)
damma_cor(fullness_table,output="table")
}
