% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_index.R
\name{damma_index}
\alias{damma_index}
\title{Aggregates raw fullness values into meaningful indices at the compound level}
\usage{
damma_index(MCI_table, pathway_table)
}
\arguments{
\item{MCI_table}{Function distillation table outputted by damma() or damma_correction}

\item{functions_table}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}

\item{weighed}{Whether to weigh average MCI values according to their hierarchical classification}
}
\value{
A MCI value for each genome or community
}
\description{
Aggregates raw fullness values into meaningful indices at the compound level
}
\examples{
damma_compounds(compound_table)
}
