% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_definition.R
\name{decompose_definition}
\alias{decompose_definition}
\title{Decomposes a definition string of a metabolic pathway/module into its information units}
\usage{
decompose_definition(definition)
}
\arguments{
\item{Definition}{Definition string of a given metabolic pathway/module}
}
\value{
A decomoposed definition string
}
\description{
Decomposes a definition string of a metabolic pathway/module into its information units
}
\examples{
decompose_definition(definition)
decompose_definition("K01580 (K13524,K07250,K00823,K16871) (K00135,K00139,K17761)")
}
