% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distillate_definition.R
\name{distillate_definition}
\alias{distillate_definition}
\title{Calculates the fullness of each hierarchical level of a metabolic pathway/module}
\usage{
distillate_definition(definition, def_table, level, expression_vector)
}
\arguments{
\item{definition}{Definition string of a given metabolic pathway/module}

\item{def_table}{Decomposed hierarchy matrix produced by create_step_matrix.R}

\item{level}{Hierarchical level}

\item{expression_vector}{Vector of expression values of functional units present in the genome}
}
\value{
A (partially) distilled definition string
}
\description{
Calculates the fullness of each hierarchical level of a metabolic pathway/module
}
\examples{
distillate_definition(definition, def_table, level, present)
}
