% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distillate_definition_community.R
\name{distillate_definition_community}
\alias{distillate_definition_community}
\title{Generates the scores of each hierarchical level of a metabolic pathway required to calculate community-level MCIs}
\usage{
distillate_definition_community(definition, def_table, level, abundance_vector)
}
\arguments{
\item{definition}{Definition string of a given metabolic pathway/module}

\item{def_table}{Decomposed hierarchy matrix produced by create_step_matrix.R}

\item{level}{Hierarchical level}

\item{abundance_vector}{Vector containing relative abundance data per identifier}
}
\value{
A (partially) distilled definition string
}
\description{
Generates the scores of each hierarchical level of a metabolic pathway required to calculate community-level MCIs
}
\examples{
distillate_definition(definition, def_table, level, present)
}
