% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_table.R
\docType{data}
\name{function_table}
\alias{function_table}
\title{Reference table containing function descriptions and metadata}
\format{
A dataframe.
}
\usage{
data(function_table)
}
\description{
Information compiled from databases KEGG, MetaSys and MEROPS.
}
\details{
Data created from version 20220619
}
\examples{
data(function_table)
}
\keyword{datasets}
