% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_matrix_list.R
\name{sweep_matrix_list}
\alias{sweep_matrix_list}
\title{Sweeps list names and row names in a list of matrices}
\usage{
sweep_matrix_list(matrix_list)
}
\arguments{
\item{matrix_list}{A list of matrices (e.g. distilled_expression_table yielded by damma_expression())}
}
\value{
A list of matrices with sweeped list names and row names
}
\description{
Sweeps list names and row names in a list of matrices
}
\examples{
sweep_matrix_list(matrix_list)
}
