% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma.R
\name{damma}
\alias{damma}
\title{Generates a gene presence-based Metabolic Capacity Index (MCI) table from a bacterial genome annotation table}
\usage{
damma(annotation_table, pathway_table, genomecol, keggcol, eccol, pepcol)
}
\arguments{
\item{annotation_table}{Table containing Genome identifiers and gene annotations}

\item{pathway_table}{Table containing definitions and metadata of metabolic functions (provided by DAMMA)}

\item{genomecol}{Column index (number) of the annotation_table containing the genome identifiers}

\item{keggcol}{Column index(es) of the annotation_table in which to search for KEGG KO annotations}

\item{eccol}{Column index(es) of the annotation_table in which to search for Enzyme Commision (EC) annotations}

\item{pepcol}{Column index(es) of the annotation_table in which to search for Peptidase annotations}
}
\value{
A pathway-level MCI table
}
\description{
Generates a gene presence-based Metabolic Capacity Index (MCI) table from a bacterial genome annotation table
}
\examples{
damma(annotation_table,pathway_table,genomecol,keggcol,eccol,pepcol)
damma(annotation_table,pathway_table,genomecol=2,keggcol=9,eccol=c(10,19),pepcol=12)
}
