% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_bin.R
\name{damma_bin}
\alias{damma_bin}
\title{Convert the fullness table of metabolic pathways/modules into a binary (present/absent) matrix}
\usage{
damma_bin(fullness_table, threshold = 0.9, completeness)
}
\arguments{
\item{fullness_table}{Table containing MAG identifiers and annotation codes}

\item{threshold}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}

\item{completeness}{Column index (number) of the annotations table containing the MAG identifiers}
}
\value{
A binary function matrix
}
\description{
Convert the fullness table of metabolic pathways/modules into a binary (present/absent) matrix
}
\examples{
damma_bin(fullness_table)
damma_bin(fullness_table,0.9)
damma_bin(fullness_table,threshold=0.9,completeness)
}
