% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_bin.R
\name{damma_bin}
\alias{damma_bin}
\title{Convert the fullness table of metabolic pathways/modules into a binary (present/absent) matrix}
\usage{
damma_bin(fullness_table, threshold = 0.9, completeness)
}
\arguments{
\item{fullness_table}{Table containing fullness scores of functions (X axis) per MAG (Y axis)}

\item{threshold}{Minimum fullness value to consider a function (pathway/module) to be present in a MAG}

\item{completeness}{Two-column table indicating the completeness (eg. CheckM) percentage (column 2) of each MAG (column 1)}
}
\value{
A binary function matrix
}
\description{
Convert the fullness table of metabolic pathways/modules into a binary (present/absent) matrix
}
\examples{
damma_bin(fullness_table)
damma_bin(fullness_table,0.9)
damma_bin(fullness_table,threshold=0.9,completeness)
}
