% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_community.R
\name{damma_community}
\alias{damma_community}
\title{Generate the fullness table of metabolic pathways/modules from a MAG annotation table}
\usage{
damma_community(
  annotations,
  functions_table,
  abundance_table,
  fullness_table,
  genomecol,
  keggcol,
  eccol,
  pepcol
)
}
\arguments{
\item{annotations}{Table containing Genome identifiers and annotation codes}

\item{functions_table}{Table containing definitions and metadata of metabolic functions (included in DAMMA package)}

\item{abundance_table}{(Relative) abundance table with samples in columns and Genomes in rows}

\item{fullness_table}{Pathway fullness table provided by damma() function}

\item{genomecol}{Column index (number) of the annotations table containing the MAG identifiers}

\item{keggcol}{Column index(es) of the annotations table in which to search for KEGG KO annotations}

\item{eccol}{Column index(es) of the annotations table in which to search for Enzyme Commision (EC) annotations}

\item{pepcol}{Column index(es) of the annotations table in which to search for Peptidase annotations}
}
\value{
A fullness matrix
}
\description{
Generate the fullness table of metabolic pathways/modules from a MAG annotation table
}
\examples{
damma_community(fullness_table,abundance_table,functions_table)
}
