% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_compounds.R
\name{damma_compounds}
\alias{damma_compounds}
\title{Aggregates raw fullness values into meaningful indices at the compound level}
\usage{
damma_compounds(distilled_table, functions)
}
\arguments{
\item{distilled_table}{Function distillation table outputted by damma()}

\item{functions}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}
}
\value{
A function table aggregated at the compound level
}
\description{
Aggregates raw fullness values into meaningful indices at the compound level
}
\examples{
damma_compounds(compound_table)
}
