% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damma_correction.R, R/damma_correlation.R
\name{damma_correlation}
\alias{damma_correlation}
\title{Models the relationship between function fullness and MAG completeness and applies a correction factor}
\usage{
damma_correlation(fullness_table, method = "spearman", output = "plot")

damma_correlation(fullness_table, method = "spearman", output = "plot")
}
\arguments{
\item{fullness_table}{Table containing fullness scores of functions (X axis) per MAG (Y axis)}

\item{method}{Correlation method to be used. Default="pearson"}

\item{output}{Whether to output a table or a plot. Default="plot"}

\item{mag_completeness}{A matrix containing MAG names (1st column) and completemess values ()}
}
\value{
A modified fullness table

A correlation matrix or plot
}
\description{
Models the relationship between function fullness and MAG completeness and applies a correction factor

Calculate correlations between compounds and core functions
}
\examples{
damma_correlation(fullness_table, mag_completeness)
damma_cor(fullness_table)
damma_cor(fullness_table,output="table")
}
