% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dama_functions.R
\name{damma_functions}
\alias{damma_functions}
\title{Aggregates compound-level MCIs into function-level MCIs}
\usage{
damma_functions(MCI_table, pathway_table)
}
\arguments{
\item{MCI_table}{Function distillation table outputted by damma() or damma_correction}

\item{pathway_table}{Table containing definitions and metadata of metabolic functions (provided by GAMMA)}

\item{transform}{Whether to transform to 0-1 scale. Default=TRUE}
}
\value{
A MCI table aggregated at the compound level
}
\description{
Aggregates compound-level MCIs into function-level MCIs
}
\examples{
dama_functions(compound_table)
}
